refactor: extract commands.rs from backend/mod.rs
Reduces mod.rs from 975 to 880 lines.
Part of upstream codebase refactoring (commit 98d215d).
914 tests pass.
